+++ api/src/Infrastructure/Model/OAuth/Entity/AccessTokenRepository.php
@@ -1,0 +1,72 @@
+<?php
+declare(strict_types=1);
+
+namespace Api\Infrastructure\Model\OAuth\Entity;
+
+use Api\Model\OAuth\Entity\AccessTokenEntity;
+use Doctrine\ORM\EntityManagerInterface;
+use Doctrine\ORM\EntityRepository;
+use League\OAuth2\Server\Entities\AccessTokenEntityInterface;
+use League\OAuth2\Server\Entities\ClientEntityInterface;
+use League\OAuth2\Server\Exception\UniqueTokenIdentifierConstraintViolationException;
+use League\OAuth2\Server\Repositories\AccessTokenRepositoryInterface;
+
+class AccessTokenRepository implements AccessTokenRepositoryInterface
+{
+    /** @var EntityRepository */
+    private $repository;
+    /** @var EntityManagerInterface */
+    private $em;
+
+    public function __construct(EntityManagerInterface $em)
+    {
+        $this->repository = $em->getRepository(AccessTokenEntity::class);
+        $this->em = $em;
+    }
+
+    public function persistNewAccessToken(AccessTokenEntityInterface $accessTokenEntity): void
+    {
+        if ($this->exists($accessTokenEntity->getIdentifier())) {
+            throw UniqueTokenIdentifierConstraintViolationException::create();
+        }
+        $this->em->persist($accessTokenEntity);
+        $this->em->flush();
+    }
+
+    public function revokeAccessToken($tokenId): void
+    {
+        if ($token = $this->repository->find($tokenId)) {
+            $this->em->remove($token);
+            $this->em->flush();
+        }
+    }
+
+    public function isAccessTokenRevoked($tokenId): bool
+    {
+        return !$this->exists($tokenId);
+    }
+
+    public function getNewToken(
+        ClientEntityInterface $clientEntity,
+        array $scopes,
+        $userIdentifier = null
+    ): AccessTokenEntityInterface {
+        $accessToken = new AccessTokenEntity();
+        $accessToken->setClient($clientEntity);
+        foreach ($scopes as $scope) {
+            $accessToken->addScope($scope);
+        }
+        $accessToken->setUserIdentifier($userIdentifier);
+
+        return $accessToken;
+    }
+
+    private function exists($id): bool
+    {
+        return $this->repository->createQueryBuilder('t')
+            ->select('COUNT(t.identifier)')
+            ->andWhere('t.identifier = :identifier')
+            ->setParameter(':identifier', $id)
+            ->getQuery()->getSingleScalarResult() > 0;
+    }
+}