+++ api/src/Infrastructure/Model/OAuth/Entity/AuthCodeRepository.php
@@ -1,0 +1,61 @@
+<?php
+declare(strict_types=1);
+
+namespace Api\Infrastructure\Model\OAuth\Entity;
+
+use Api\Model\OAuth\Entity\AuthCodeEntity;
+use Doctrine\ORM\EntityManagerInterface;
+use Doctrine\ORM\EntityRepository;
+use League\OAuth2\Server\Entities\AuthCodeEntityInterface;
+use League\OAuth2\Server\Exception\UniqueTokenIdentifierConstraintViolationException;
+use League\OAuth2\Server\Repositories\AuthCodeRepositoryInterface;
+
+class AuthCodeRepository implements AuthCodeRepositoryInterface
+{
+    /** @var EntityRepository */
+    private $repository;
+    /** @var EntityManagerInterface */
+    private $em;
+
+    public function __construct(EntityManagerInterface $em)
+    {
+        $this->repository = $em->getRepository(AuthCodeEntity::class);
+        $this->em = $em;
+    }
+
+    public function persistNewAuthCode(AuthCodeEntityInterface $authCodeEntity): void
+    {
+        if ($this->exists($authCodeEntity->getIdentifier())) {
+            throw UniqueTokenIdentifierConstraintViolationException::create();
+        }
+        $this->em->persist($authCodeEntity);
+        $this->em->flush();
+    }
+
+    public function revokeAuthCode($codeId): void
+    {
+        if ($code = $this->repository->find($codeId)) {
+            $this->em->remove($code);
+            $this->em->flush();
+        }
+    }
+
+    public function isAuthCodeRevoked($codeId): bool
+    {
+        return !$this->exists($codeId);
+    }
+
+    public function getNewAuthCode(): AuthCodeEntity
+    {
+        return new AuthCodeEntity();
+    }
+
+    private function exists($id): bool
+    {
+        return $this->repository->createQueryBuilder('t')
+            ->select('COUNT(t.identifier)')
+            ->andWhere('t.identifier = :identifier')
+            ->setParameter(':identifier', $id)
+            ->getQuery()->getSingleScalarResult() > 0;
+    }
+}