+++ api/src/Model/OAuth/Entity/ScopeEntity.php
@@ -1,0 +1,18 @@
+<?php
+declare(strict_types=1);
+
+namespace Api\Model\OAuth\Entity;
+
+use League\OAuth2\Server\Entities\ScopeEntityInterface;
+use League\OAuth2\Server\Entities\Traits\EntityTrait;
+use League\OAuth2\Server\Entities\Traits\ScopeTrait;
+
+class ScopeEntity implements ScopeEntityInterface
+{
+    use EntityTrait, ScopeTrait;
+
+    public function __construct(string $identifier)
+    {
+        $this->identifier = $identifier;
+    }
+}